refactor: move schema inference module to helpers
